requirements.txt
discord.py>=2.3.0
httpx>=0.27.0
feedparser>=6.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
click>=8.1.0
pytz>=2024.1
